Move mock-server into standalone directory
